Feature/mw 477 add remote config to android app (#67)
* Upgraded gradle dependencies to accommodate ML

* Added initial ImageEstimator Classes

* Fixed gradle conflicts

* Removed Sceneform Dependencies

* Converted to TensorFlow Lite

* Added MLModel input handling

* Improved YUV to RGB conversion performance

* Added score calculations and improve RGB value gathering performance

* Added ImageQuality frame filter

* Added renderscript deallocation and updated filtering methods

* Added GPU acceleration

* Converted ImageQualityEstimator Class to FrameFilter

* Reestablished BlurFilter

* Added ImageQualityFilter Statistics

* Fixed race condition on YUV to RGB conversion

* Added ImageQuality statistics to StatisticsView

* Updated aar file

* Added remote model GET method

* Improve Storage reading performance

* Added asset default-config.json parsing

* Fixed camera pitch filter

* Updated imageQualityFilter

* Added default config file

* Removed isZoneInRadius Request

* Added remote config request

* Fixed Remote Model retrieving

* Fixed ImageQualityModel input and output calculations

* Added Accuracy to initializeRequest

* Updated permissions

* Added Location validation on localize call

* Added multithread support to model interpreter

* Added GPU support

* Added deviceOs tag on initializeRequest

* Updated documentation

* Removed modelBinding

+++ FantasmoSDK/app/src/main/java/com/fantasmo/sdk/models/analytics/AccumulatedARCoreInfo.kt
@@ -10,7 +10,7 @@ import com.google.ar.core.Frame
 class AccumulatedARCoreInfo {
 
     var modelVersion: String = ""
-    var scoreThreshold: Double = 0.0
+    var scoreThreshold: Float = 0f
     var lastImageQualityScore: Float = 0f
     var trackingStateFrameStatistics = TrackingStateFrameStatistics()
     var elapsedFrames = 0
@@ -43,7 +43,7 @@ class AccumulatedARCoreInfo {
         translationAccumulator.reset()
         rotationAccumulator.reset()
         modelVersion = ""
-        scoreThreshold = 0.0
+        scoreThreshold = 0f
         lastImageQualityScore = 0f
     }
 }
